refactor(test): move test constants from src to test directory
