+++ src/components/BoardControls/utils/sortByLikes.test.ts
@@ -1,4 +1,4 @@
-import { USER_TEST_ID as userId } from '../../../constants/test';
+import { USER_TEST_ID as userId } from '../../../../test/constants';
 import { sortByLikes } from './sortByLikes';
 
 const column = {
