+++ src/components/Item/Likes.test.tsx
@@ -1,10 +1,10 @@
 import { fireEvent, screen } from '@testing-library/react';
 
-import { renderWithProviders, updateStore } from '../../../test/utils';
 import {
   BOARD_TEST_ID as boardId,
   ITEM_TEST_ID as itemId,
-} from '../../constants/test';
+} from '../../../test/constants';
+import { renderWithProviders, updateStore } from '../../../test/utils';
 import Likes from './Likes';
 
 const props = {
