+++ src/store/slices/boardsSlice.test.ts
@@ -1,7 +1,7 @@
 import {
   BOARD_TEST_ID as boardId,
   USER_TEST_ID as userId,
-} from '../../constants/test';
+} from '../../../test/constants';
 import { boardsSlice, initialState } from './boardsSlice';
 
 const { actions, reducer } = boardsSlice;
