+++ src/store/slices/likesSlice.test.ts
@@ -2,7 +2,7 @@ import {
   BOARD_TEST_ID as boardId,
   ITEM_TEST_ID as itemId,
   USER_TEST_ID as userId,
-} from '../../constants/test';
+} from '../../../test/constants';
 import { initialState, likesSlice } from './likesSlice';
 
 const { actions, reducer } = likesSlice;
